Fix incorrect remapping of tracks from cache
There are some situations where a given Spotify track can match with
multiple Tidal tracks, or multiple Spotify tracks can match with a
single Tidal track. For example when you have duplicate tracks on the
Spotify playlist which are from single vs. album version so they have
different spotify IDs.

For example we might have the following two cases for Spotify/Tidal
Case 1:                           Case 2:
A  A' (matches both A and C)      A  A' (matches both A and C)
B  D'                             B  C' (matches both A and C)
C  (matched A')                   C  D'
D                                 D

The last algorithm failed for case 2 because it would remap C to A'
and so C' would get deleted, which significantly slows the script down.

The new algorithm handles both cases so long as there was no reordering
occurring. In the case where there *was* reordering it would just mean
that A' and C' could get their order swapped which is a very small
problem and the speed slowdown would have occurred anyway due to
reordering.